Create a separate sso_client
